pulse-generator: Fix period/buffer size errors.
The pulse-generator intends to write one pulse for each period.
However, it was mixing up frames, bytes, periods, and buffer sizes.
E.g. if the period size was 160 frames and the buffer size was 480
frames... it was trying to write 480 bytes every time (but the period
size is 640 for 2-chan. 16-bit audio).

Other changes in this commit:

   * Set the threshhold and stop configuration to get an interrupt
     at the period boundary.

   * Added support for putting the pulse in the middle of the period.

   * Changed pulse byte from 0x7f to 0x1f (-13 dB), which is a bit
     more comfortable to listen to.

   * Removed the logging inside ivt_lap().

Change-Id: I956ec0d06f125fca5fb012fa7c402947d22c0ff8
Signed-off-by: Gabriel M. Beddingfield <gabrbedd@ti.com>